nvim0: VeryLazy instead of InsertEnter and BufReadPre
I rather wait a bit longer initially instead of when loading the first
buffer or inserting something since that will be the first things I will
do anyway. It does not feel right when opening a buffer is delayed, even
when it is just the first time. Even worse for inserting